requirements.txt
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
huggingface-hub==0.25.1
hydra-core==1.3.2
idna==3.8
imageio==2.35.1
importlib_metadata==8.2.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.3
numpy==2.0.1
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyDOE==0.3.8
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.0
seaborn==0.13.2
setuptools==72.1.0
six==1.16.0
stack-data==0.6.2
sympy==1.13.2
threadpoolctl==3.5.0
timm==1.0.9
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
wheel==0.44.0
zipp==3.19.2
